Only reserve a single port
Changed the document to only reserve a single port. RFC 3551 reserved
two ports for RTP A/V streams, so there is no longer a use case for
reserving multiple ports. Renames the document accordingly (resetting
the version number) and example code files.

Removed reference to RFC 3550, as it is no longer needed.

Added wording to indicate that the requested port is only REQUESTED,
not assigned yet.

Changed the requested port to 49149 to avoid conflicts with existing
reservations at 49151 and 49150.