Change strategy for non-conformant host compatibility
The previous strategy for updating hosts was to require hosts to
regard any use of this port as non-exclusive. The advantage of this
is that as hosts are updated the SO_REUSEADDR/PORT socket options
are no longer necessary.

The disadvantage is that an application developed on a conformant
host will no longer work when running on a non-conformant host.
Even worse, this application could claim exclusive use of the port,
preventing other applications running on that host from binding to
the socket.

Requiring applications on conformant hosts to set the appropriate
socket options will ensure those applications function properly
on non-conformant hosts.